more-scenarios/medical/requirements.txt
absl-py==2.0.0
addict==2.4.0
attrs==23.1.0
aws-requests-auth==0.4.3
batchgenerators==0.25
boto3==1.28.62
botocore==1.31.62
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
connected-components-3d==3.12.3
contourpy==1.1.1
cucim==23.8.0
cupy-cuda11x==12.2.0
cycler==0.11.0
fastremap==1.14.0
fastrlock==0.8.2
fonttools==4.42.1
future==0.18.3
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h5py==3.10.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
jmespath==1.0.1
joblib==1.3.2
jsonlines==1.2.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
linecache2==1.0.0
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
monai==1.2.0
networkx==3.1
nibabel==5.1.0
numpy==1.24.4
nvidia-ml-py==12.535.108
nvitop==1.3.0
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prefetch-generator==1.0.3
prettytable==3.9.0
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.4.3
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
quilt3==5.3.1
referencing==0.30.2
requests==2.31.0
requests-futures==1.0.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
rsa==4.9
s3transfer==0.7.0
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
SimpleITK==2.0.2
six==1.16.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
traceback2==1.4.0
typing_extensions==4.8.0
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
xlrd==2.0.1
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
